requirements.txt
Django>=5.0.1
djangorestframework>=3.14.0
python-dotenv>=1.0.0
celery>=5.3.6
redis>=5.0.1

# Web scraping
firecrawl-py>=0.0.16
beautifulsoup4>=4.12.3
lxml>=5.1.0
requests>=2.31.0

# PDF generation
reportlab>=4.0.9
# weasyprint - Optional, can cause install issues on some systems

# AWS S3 / Cloudflare R2
boto3>=1.34.34

# LLM Integration
groq>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Utils
python-dateutil>=2.8.2
Pillow>=10.4.0
